trasn-jimlj/backend/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bitarray==2.9.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
ckzg==1.0.2
constantly==23.10.4
cryptography==42.0.5
cytoolz==0.12.3
daphne==4.1.2
Django==5.0.7
django-cors-headers==4.3.1
django-formtools==2.5.1
django-friendship==1.9.6
django-otp==1.5.4
django-phonenumber-field==7.3.0
django-seed==0.3.1
django-seeding==1.2.5
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==5.0.0
eth-utils==4.1.1
eth_abi==5.1.0
Faker==30.0.0
frozenlist==1.4.1
gunicorn==22.0.0
hexbytes==0.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
msgpack==1.1.0
multidict==6.1.0
packaging==23.2
parsimonious==0.10.0
pillow==10.4.0
protobuf==5.28.2
psycopg2-binary==2.9.9
py-solc-x==2.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pyunormalize==16.0.0
qrcode==7.4.2
redis==5.0.7
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
rlp==4.0.1
rpds-py==0.20.0
service-identity==24.1.0
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
toolz==0.12.1
toposort==1.10
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.1
web3==6.20.0
websockets==10.3
wheel==0.43.0
yarl==1.13.0
zope.interface==6.3
